requirements.txt
alembic==1.11.2
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2==2.9.7
PyJWT==2.8.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
